requirements.txt
Django>=4.2
djangorestframework>=3.14.0
drf-yasg>=1.21.7
numpy>=1.24.4
pandas>=2.0.3
scikit-learn>=1.3.0
joblib>=1.3.2